requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
python-dotenv==0.19.0
supabase==0.7.1 